Move Features.Report namespace to Features.Reports (pluralized)
