+++ src/Nyss.Web/Features/HealthRisks/HealthRiskServicesRegistration.cs
@@ -6,7 +6,7 @@ namespace Nyss.Web.Features.HealthRisks
     {
         public static void RegisterHealthRiskFeature(this IServiceCollection services)
         {
-            services.AddScoped<IHealthRisksService, HealthRisksService>();
+            services.AddSingleton<IHealthRisksService, HealthRisksService>();
         }
     }
 }
