+++ src/Nyss.Web/Features/Reports/Export/Formats/ExcelExporter.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using OfficeOpenXml;
 
-namespace Nyss.Web.Features.Report.Export.Formats
+namespace Nyss.Web.Features.Reports.Export.Formats
 {
     public class ExcelExporter : ICanExportCaseReports
     {
